// src/types.ts
export interface NatsMessage {
	/**
	 * Defines the message's subject.
	 * @type {string}
	 */
	subject: string;

	/**
	 * The message itself.
	 * @type {string}
	 */
	message: Uint8Array;

	/**
	 * Optional set of key/value associated with the message.
	 * @type {NatsMessageHeaders | undefined}
	 */
	headers?: NatsMessageHeaders;
};

/**
 * Contents of a presistnt message.
 * @interface NatsStreamMessage
 */
export interface NatsStreamMessage extends NatsMessage {
	/**
	 * The sequence number of the message in the stream.
	 * @type {number}
	 */
	sequence: number;

	/**
	 * The time the message was received.
	 * @type {Date}
	 */
	timestamp: Date;
};

/**
 * A persistent message that must be acknowledged.
 * @interface NatsStreamMessageAck
 */
export interface NatsStreamMessageAck extends NatsStreamMessage {
	/**
	 * Call this method to acknowledge reception and handling of this message.
	 * @method ack
	 */
	ack: () => void

	/**
	 * Indicate to the JetStream server that processing of the message failed,
	 * and that it should be resent after the spefied number of milliseconds.
	 * @param {number|undefined} millis
	 */
	nak: (millis?: number) => void
};
// ...
export interface NatsClientConsumer {
	/**
	 * Instructs the server to delete this consumer.
	 * @async
	 * @method delete
	 */
	delete: () => Promise<void>;

	/**
	 * Starts a subscription for messages based on the given stream and consumer.
	 * @async
	 * @method subscribe
	 * @param {NatsStreamMessageCallback} cb - Asynchronous callback to call when a new message arrives.
	 */
	subscribe: (cb: NatsStreamMessageCallback) => Promise<void>;

	/**
	 * Ends the subscription.
	 * It does not throw errors if the connection is closed or the subscription does not exists.
	 * @async
	 * @method subscribe
	 */
	unsubscribe: () => Promise<void>;
};
// ...
export type NatsStreamMessageCallback = (msg: NatsStreamMessageAck) => Promise<boolean | undefined>;